Prueba exitosa con el binding de Haskell

+++ EsferixisMusicFrameworkNative/EMFB_STK/include/EMFB_STK_FRAMES.h
@@ -4,7 +4,7 @@
 
 EMFB_STK_API void * emfb_stk_stkframes_new_zero(char **exception_desc, unsigned int nFrames, unsigned int nChannels);
 
-EMFB_STK_API void * emfb_stk_stkframes_new_valued(char **exception_desc, float value, unsigned int nFrames, unsigned int nChannels);
+EMFB_STK_API void * emfb_stk_stkframes_new_valued(char **exception_desc, double value, unsigned int nFrames, unsigned int nChannels);
 
 EMFB_STK_API unsigned int emfb_stk_stkframes_channels(void *frames);
